requirements.txt
requests>=2.31.0
packaging>=23.1
semver>=3.0.0
PyYAML>=6.0
python-dateutil>=2.8.2
pytest>=7.0.0
pytest-mock>=3.10.0
